Don't look back in angohr...
Initial commit looking at using the latest CMake with the latest C++
feature I have been wanting to play with - modules. Making use of
CMakePresets.json to set initial values for a few things like the
generator, vcpkg, etc.

This uses C++ 20 for a couple of targets, and then makes use of an
imported module to do the traditional hello world type thing. Just
scratching the surface on the whole module thing clearly, but wanted to
make sure it even worked locally as there are very few examples.